created a basic "quick buy" and a middleman
the quick buy is very rudimentary and requires more work, but it can be used to help understand how the code is supposed to look. It requires a bit of polishing to follow the exact logic we want and a bit of a clean up. Middleman is supposed to help fetch data from supabase, currently it only fetches profiles and Portfolio.